Geom vs ScatterDraw nameclash fixed
git-svn-id: svn://ultimatepp.org/upp/trunk@15120 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Util.h
@@ -479,7 +479,7 @@ size_t SizeBySerialize(const T& cont)
 	TimeStop tm;
 	SizeStream szs;
 	const_cast<T&>(cont).Serialize(szs);
-	return szs;
+	return (size_t)szs;
 }
 
 template <class T>
